requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-dotenv
passlib[bcrypt]
python-jose
python-multipart
Pillow
email-validator